Create vertical_polar_forcing
This is data for Atmospheric circulation sensitivity to changes in the vertical structure of polar warming. Please refer to the method section.
The name of GRAM experiments, bot, mid1, mid2, and mid3 indicate L990, L850, L700, and L550 forcing respectively.
Depending on the forcing amplitude, it varies 1X to 5.5X. The control run for GRAM is ctl_GRAM.nc.
There are additional data for L990 and L650 for AM2 experiments. The control run for AM2 is ctl_AM2.nc.